+++ CSRedis/Internal/IO/IRedisSocket.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Net;
+using System.Net.Sockets;
+using System.Text;
+
+namespace CSRedis.Internal.IO
+{
+    interface IRedisSocket : IDisposable
+    {
+        bool Connected { get; }
+        int ReceiveTimeout { get; set; }
+        int SendTimeout { get; set; }
+        void Connect(EndPoint endpoint);
+        bool ConnectAsync(SocketAsyncEventArgs args);
+        bool SendAsync(SocketAsyncEventArgs args);
+        Stream GetStream();
+    }
+}
